Add more checks to deleteContact
Summary: Deleting files without checking whether the path is in our vcards location or whether this data source is responsible at all is probably a bad idea.

Test Plan: Deleting still works in plasma-phonebook

Reviewers: apol

Reviewed By: apol

Differential Revision: https://phabricator.kde.org/D23933

+++ src/kpeoplevcard.cpp
@@ -133,8 +133,15 @@ bool VCardDataSource::addContact(const QVariantMap & properties)
 
 bool VCardDataSource::deleteContact(const QString &uri)
 {
+    if (!uri.startsWith("vcard:/"))
+        return false;
+
     QString path = uri;
     path.remove("vcard:/");
+
+    if (!path.startsWith(*vcardsLocation))
+        return false;
+
     return QFile::remove(path);
 }
